group member base routes

+++ src/WebApi/Controllers/GroupMemberController.cs
@@ -1,11 +1,12 @@
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using WebApi.Context.Interfaces;
-using WebApi.DTOs.Requests;
+using WebApi.DTOs.Requests.GroupMember;
 using WebApi.DTOs.Responses;
-using WebApi.Models;
+using WebApi.DTOs.Responses.GroupMember;
 using WebApi.Persistence.Interfaces;
 using WebApi.Repositories.Interfaces;
+using WebApi.Security;
 
 namespace WebApi.Controllers;
 
@@ -19,10 +20,55 @@ public class GroupMemberController(
     IGroupRepository groupRepository,
     IGroupMemberRepository groupMemberRepository) : ControllerBase
 {
+    [HttpGet]
+    [Authorize(Policy = Policies.Admin)]
+    public async Task<IActionResult> GetAllGroupMembers([FromRoute] Guid groupId, 
+        [FromQuery] int pageNumber = 1, [FromQuery] int pageSize = 10)
+    {
+        var (groupMembers, hits) = 
+            await groupMemberRepository.GetAllByGroupIdAsync(groupId, pageNumber, pageSize);
+
+        return Ok(new GetAllGroupMembersResponse
+        {
+            Data = groupMembers.Select(x => new GetAllGroupMembersResponseData
+            {
+                Id = x.Id,
+                CreatedAt = x.CreatedAt
+            }),
+            Hits = hits
+        });
+    }
+
+    [HttpGet("{id:guid}")]
+    [Authorize(Policy = Policies.Admin)]
+    public async Task<IActionResult> GetGroupMemberById([FromRoute] Guid groupId, [FromRoute] Guid id)
+    {
+        var group = await groupRepository.GetGroupWithMembersAsync(groupId);
+
+        if (group is null)
+        {
+            return NotFound(new ErrorResponse("Group not found."));
+        }
+
+        var groupMember = group.Members.SingleOrDefault(x => x.Id == id);
+
+        if (groupMember is null)
+        {
+            return NotFound(new ErrorResponse("Group member not found."));
+        }
+
+        return Ok(new GetGroupMemberByIdResponse
+        {
+            UserId = groupMember.UserId,
+            IsOwner = groupMember.IsOwner,
+            CreatedAt = groupMember.CreatedAt,
+            UpdatedAt = groupMember.UpdatedAt
+        });
+    }
+
     [HttpDelete]
-    public async Task<IActionResult> RemoveUserFromGroup(
-        [FromRoute] Guid groupId, 
-        [FromBody] RemoveUserFromGroupRequest request)
+    public async Task<IActionResult> RemoveGroupMemberById([FromRoute] Guid groupId,
+        [FromBody] RemoveGroupMemberByIdRequest request)
     {
         var currentUserId = currentUser.GetId();
 
@@ -32,12 +78,14 @@ public class GroupMemberController(
         }
 
         var group = await groupRepository.GetGroupWithMembersAsync(groupId);
+
         if (group is null)
         {
             return NotFound(new ErrorResponse("Group not found."));
         }
 
         var currentUserMember = group.Members.SingleOrDefault(m => m.UserId == currentUserId);
+
         if (currentUserMember is null)
         {
             return BadRequest(new ErrorResponse("You are not a member of this group."));
@@ -48,10 +96,16 @@ public class GroupMemberController(
             return BadRequest(new ErrorResponse("Only the group owner can remove members."));
         }
 
-        var userToRemove = group.Members.SingleOrDefault(m => m.UserId == request.UserId);
+        var userToRemove = group.Members.SingleOrDefault(m => m.Id == request.MemberId);
+
         if (userToRemove is null)
         {
             return NotFound(new ErrorResponse("The user is not a member of this group."));
+        }
+
+        if (currentUser.GetId() == userToRemove.UserId)
+        {
+            return BadRequest(new ErrorResponse("User can not remove itself."));
         }
 
         groupMemberRepository.Remove(userToRemove);
